requirements.txt
pandas==2.3.0
numpy==2.2.6
haversine==2.9.0
plotly==6.1.2
streamlit==1.46.0
folium==0.20.0
streamlit_folium==0.25.0
Pillow==11.2.1